Move ExpandableNumberInput to its own component - Continuous number mapper uses ExpandableNumberInput instead of TextFields, allowing users to cancel/confirm new number values alongside min/max validation - Various style tweaks for the color mapper and number mapper dialog.
